test: S12.21 migrate message-formatter field tests to unit file
Move the PRIVAL, hostname/app-name/proc-id, msgid and all-fields tests into a
new SolidSyslogMessageFormatterTest.cpp that drives SolidSyslogMessageFormatter_
Format directly (struct message + context + output buffer parsed via
SyslogFieldParser). The three NullGet*ProducesNilvalue tests stay in
SolidSyslogTest.cpp — like S12.20's NullClock split, they assert Create's
nil-object substitution, not formatter behaviour; the formatter's empty->NILVALUE
path is covered by Empty*ProducesNilvalue in the unit file. SD/MSG-body/buffer/
Service/lifecycle tests stay as integration coverage. Drop the now-dead
CHECK_PRIVAL/CHECK_MSGID macros and TEST_PRIVAL/TEST_MSGID constants.
